requirements.txt
azure_storage==0.37.0
fastapi==0.115.5
langchain==0.3.7
langchain_community==0.3.7
langchain_core==0.3.17
langchain_openai==0.2.8
langchain_pinecone==0.2.0
langgraph==0.2.46
pydantic==2.9.2
python-dotenv==1.0.1
typing_extensions==4.12.2
